set new rating

+++ data_remote/src/main/java/msk/pobazar/wcquiz/data_remote/models/RatingResponse.kt
@@ -6,7 +6,7 @@ data class RatingResponse(
     val countAll: Int? = 0,
     val countRight: Int? = 0,
     val score: Float? = 0F,
-    val time: Float? = 0F,
+    val time: Long? = 0L,
     val user: String? = "",
     val winStrick: Int? = 0,
     val date: Date? = Date(0)
