+++ domain/src/main/java/msk/pobazar/wcquiz/domain/model/Rating.kt
@@ -5,7 +5,7 @@ import java.util.Date
 data class Rating(
     val countRight: Int,
     val score: Float,
-    val time: Float,
+    val time: Long,
     val user: String,
     val date: Date
 )
